raw_kafka_feature_consumer_pod/requirements.txt
# scikit-learn, pandas, numpy and scipy are included in the 'alpine-python-machinelearning' image
kafka-python
redis